# src/main/apiServer/generated/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cherry Studio API",
    "version": "1.0.0",
    "description": "OpenAI-compatible API for Cherry Studio with additional Cherry-specific endpoints",
    "contact": {
      "name": "Cherry Studio",
      "url": "https://github.com/CherryHQ/cherry-studio"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use the API key from Cherry Studio settings"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "AgentType": {
        "type": "string",
        "enum": ["claude-code"],
        "description": "Type of agent"
      },
...
      "AgentBase": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Agent name"
          },
          "description": {
            "type": "string",
            "description": "Agent description"
          },
          "accessible_paths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of directory paths the agent can access"
          },
          "instructions": {
            "type": "string",
            "description": "System prompt/instructions"
          },
          "model": {
            "type": "string",
            "description": "Main model ID"
          },
          "plan_model": {
            "type": "string",
            "description": "Optional planning model ID"
          },
          "small_model": {
            "type": "string",
            "description": "Optional small/fast model ID"
          },
          "mcps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of MCP tool IDs"
          },
          "allowed_tools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of allowed tool IDs (whitelist)"
          },
          "configuration": {
            "$ref": "#/components/schemas/AgentConfiguration"
          }
        },
        "required": ["model", "accessible_paths"]
      },
      "AgentEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AgentBase"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique agent identifier"
              },
              "type": {
                "$ref": "#/components/schemas/AgentType"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "description": "ISO timestamp of creation"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "description": "ISO timestamp of last update"
              }
            },
            "required": ["id", "type", "created_at", "updated_at"]
          }
        ]
      },
      "CreateAgentRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AgentBase"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/AgentType"
              },
              "name": {
                "type": "string",
                "minLength": 1,
                "description": "Agent name (required)"
              },
              "model": {
                "type": "string",
                "minLength": 1,
                "description": "Main model ID (required)"
              }
            },
            "required": ["type", "name", "model"]
          }
        ]
      },
...
      "ListAgentsResponse": {
        "type": "object",
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentEntity"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of agents"
          },
          "limit": {
            "type": "integer",
            "description": "Number of items returned"
          },
          "offset": {
            "type": "integer",
            "description": "Number of items skipped"
          }
        },
        "required": ["agents", "total", "limit", "offset"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "type": {
                "type": "string",
                "description": "Error type"
              },
              "code": {
                "type": "string",
                "description": "Error code"
              }
            },
            "required": ["message", "type", "code"]
          }
        },
        "required": ["error"]
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/v1/agents": {
      "post": {
        "summary": "Create a new agent",
        "description": "Creates a new autonomous agent with the specified configuration and automatically provisions an initial session that mirrors the agent's settings.",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all agents",
        "description": "Retrieves a paginated list of all agents",
        "tags": ["Agents"],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of agents to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of agents to skip"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["created_at", "updated_at", "name", "sort_order"],
              "default": "sort_order"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "description": "Sort order (asc = ascending, desc = descending). Defaults to asc when sortBy is sort_order."
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AgentEntity"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total number of agents"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Number of agents returned"
                    },
                    "offset": {
                      "type": "integer",
                      "description": "Number of agents skipped"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/agents/reorder": {
      "put": {
        "summary": "Reorder agents",
        "description": "Sets the display order of agents based on the provided array of agent IDs",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ordered_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of agent IDs in the desired display order"
                  }
                },
                "required": ["ordered_ids"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agents reordered successfully"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/agents": {
      "post": {
        "summary": "Create a new agent",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEntity"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all agents with pagination",
        "tags": ["Agents"],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of agents to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of agents to skip"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["idle", "running", "completed", "failed", "stopped"]
            },
            "description": "Filter by agent status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAgentsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}